requirements.txt
django>=5.0,<7.0
pillow>=10.0
python-decouple>=3.8
djangorestframework>=3.14
gunicorn>=21.2
psycopg2-binary>=2.9
whitenoise>=6.6
